requirements.txt

# see http://djangoproject.com
django

# see https://www.django-rest-framework.org
djangorestframework
